Fix cross-platfrom building issues
Seems like I have to specify the `kinds` globally because of cross
platform building and then handle invalidity at runtime. For example,
you would not be able to build binary on a Linux machine if you
referenced `daemon.UserAgent` since that constant would not be made
available by the compiler on a Linux build.